🧡 Move api directory to src/
