+++ next.config.js
@@ -2,10 +2,12 @@
  * @type {import('next').NextConfig}
  */
 const prefix =
-  process.env.NODE_ENV === 'production' ? 'https://isc-exhibition.github.io/ischive/' : ''
+  process.env.NODE_ENV === "production"
+    ? "https://isc-exhibition.github.io/ischive/"
+    : "";
 
 const nextConfig = {
-  output: "export",
+  // output: "export",
 
   // Optional: Change links `/me` -> `/me/` and emit `/me.html` -> `/me/index.html`
   trailingSlash: true,
